+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow encoder subclasses to customize the message before it is converted to String
   [\#40](https://github.com/osiegmar/logback-gelf/issues/40)
 - Server certificate hostname verification in `GelfTcpTlsAppender`.
+- Add server certificate whitelisting (`trustedServerCertificate`).
 
 ### Changed
 - Upgrade to Java 8 (Premier Support of Java 7 ended in July 2019)